requirements.txt
Django==5.2
djangorestframework==3.15.0
djoser==2.2.0
django-filter==24.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==20.1.0 